Cleanup P3A feature usage struct and add unit tests
